% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_methods.R,
%   R/zzz_ReshapedNiftiArray-class.R
\name{nifti_header}
\alias{nifti_header}
\alias{nifti_header,NiftiArraySeed-method}
\alias{nifti_header,NiftiArray-method}
\alias{nifti_header,NiftiArrayList-method}
\alias{nifti_header,DelayedArray-method}
\alias{nifti_header,HDF5Array-method}
\alias{nifti_header,HDF5ArraySeed-method}
\alias{nifti_header,ANY-method}
\alias{nifti_header,ReshapedNiftiArray-method}
\alias{nifti_header,ReshapedNiftiArraySeed-method}
\title{Dump a NIfTI header}
\usage{
nifti_header(image)

\S4method{nifti_header}{NiftiArray}(image)

\S4method{nifti_header}{NiftiArrayList}(image)

\S4method{nifti_header}{DelayedArray}(image)

\S4method{nifti_header}{HDF5Array}(image)

\S4method{nifti_header}{HDF5ArraySeed}(image)

\S4method{nifti_header}{ANY}(image)

\S4method{nifti_header}{NiftiArraySeed}(image)

\S4method{nifti_header}{ReshapedNiftiArray}(image)

\S4method{nifti_header}{ReshapedNiftiArraySeed}(image)
}
\arguments{
\item{image}{A file path to a NIfTI image on disk or a NiftiArray object loaded locally in R.}
}
\value{
A list of class \code{niftiHeader}, which has
the NIfTI header information.
}
\description{
Extracts the contents of a NIfTI header, from a NIfTI file on disk or a
\code{NiftiArray} object loaded into R.
}
\examples{
nii_fname = system.file("extdata",
"example.nii.gz", package = "RNifti")
nifti_header(nii_fname)
res = writeNiftiArray(nii_fname)
nifti_header(res)
}
