% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiArray.R
\name{NiftiArray}
\alias{NiftiArray}
\title{Construct NiftiArray}
\usage{
NiftiArray(filepath, name = "image", header_name = "hdr", type = NA)
}
\arguments{
\item{type}{}
}
\description{
Construct NiftiArray
}
