% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NiftiArrayList.R
\name{NiftiArrayList}
\alias{NiftiArrayList}
\title{Write NiftiArray Object}
\usage{
NiftiArrayList(x, ...)
}
\arguments{
\item{x}{a character or list of objects to be called with
\link{writeNiftiArray}}

\item{...}{additional arguments to pass to \link{writeNiftiArray}}
}
\description{
Write NiftiArray Object
}
\examples{
nii_fname = system.file("extdata", "example.nii.gz", package = "RNifti")
nii_fname = rep(nii_fname, 3)
res = NiftiArrayList(nii_fname)
testthat::expect_is(res, "NiftiArrayList")
mat = as(res, "NiftiMatrix")
arr = as(res, "NiftiArray")
}
