% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_niftiArraySeed.R
\name{NiftiArraySeed}
\alias{NiftiArraySeed}
\title{Seed for NiftyArray Class}
\usage{
NiftiArraySeed(filepath, name = "image", header_name = "hdr",
  type = NA)
}
\arguments{
\item{type}{}
}
\description{
Seed for NiftyArray Class
}
