% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_niftiArraySeed.R, R/zzz_NiftiMatrix.R
\docType{class}
\name{NiftiMatrix-class}
\alias{NiftiMatrix-class}
\alias{matrixClass}
\alias{matrixClass,NiftiArray-method}
\title{NiftiMatrix Class}
\usage{
\S4method{matrixClass}{NiftiArray}(x)
}
\arguments{
\item{x}{Typically a DelayedArray object.}
}
\description{
NiftiMatrix Class
}
