% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_ReshapedNiftiArray-class.R
\name{ReshapedNiftiArraySeed}
\alias{ReshapedNiftiArraySeed}
\alias{ReshapedNiftiArray}
\title{Virtually reshaped Nifti images as DelayedArray objects}
\usage{
ReshapedNiftiArraySeed(filepath, name = "image", header_name = "hdr",
  dim, type = NA, header = NULL)

ReshapedNiftiArray(filepath, name = "image", header_name = "hdr", dim,
  type = NA, header = NULL)
}
\arguments{
\item{filepath}{The path (as a single character string) to the HDF5
file where the dataset is located.}

\item{name}{The name of the image in the HDF5 file.}

\item{header_name}{The name of the header in the HDF5 file.}

\item{dim}{A vector of dimensions that describes the virtual
reshaping i.e. the reshaping that is virtually applied upfront
to the HDF5 dataset when the ReshapedHDF5Array object gets constructed.
See \link[HDF5Array:ReshapedHDF5Array]{HDF5Array::ReshapedHDF5Array}.}

\item{type}{\code{NA} or the R atomic type, passed to
\code{\link[HDF5Array:HDF5Array]{HDF5Array::HDF5Array()}}}

\item{header}{list of header information to override call of
\link{nifti_header}}
}
\value{
An object of class \code{ReshapedNiftiArraySeed}
}
\description{
Virtually reshaped Nifti images as DelayedArray objects
}
\examples{
nii_fname = system.file("extdata",
                        "example.nii.gz", package = "RNifti")
res = NiftiArray::NiftiArray(nii_fname)
dim(res) = c(dim(res), 1)
res = writeNiftiArray(res)
A <- ReshapedNiftiArray(
  filepath = res@seed@filepath,
  name = res@seed@name,
  header_name = res@seed@header_name,
  dim=c(prod(dim(res)[1:3]), 1))
}
