% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_methods.R
\docType{methods}
\name{nifti_header}
\alias{nifti_header}
\alias{nifti_header,NiftiArraySeed-method}
\title{Dump a NIfTI header}
\usage{
nifti_header(image)

\S4method{nifti_header}{NiftiArraySeed}(image)
}
\arguments{
\item{image}{An image or NiftiArray object.}
}
\description{
Dump a NIfTI header
}
\examples{
nii_fname = system.file("extdata", "example.nii.gz", package = "RNifti")
nifti_header(nii_fname)
res = writeNiftiArray(nii_fname)
nifti_header(res)
}
