% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_niftiArraySeed.R, R/niftiArray.R,
%   R/zzz_NiftiMatrix.R
\docType{class}
\name{NiftiArray-class}
\alias{NiftiArray-class}
\alias{NiftiArray}
\alias{class:NiftiArray}
\alias{coerce}
\alias{coerce,ANY,NiftiArray-method}
\alias{coerce,DelayedArray,NiftiArray-method}
\alias{coerce,DelayedMatrix,NiftiMatrix-method}
\alias{coerce,NiftiArray,NiftiMatrix-method}
\alias{coerce,NiftiArrayList,NiftiMatrix-method}
\alias{coerce,numeric,NiftiMatrix-method}
\alias{coerce,NiftiArray,niftiImage-method}
\alias{coerce,NiftiMatrix,niftiImage-method}
\alias{coerce,NiftiMatrix,NiftiArray-method}
\alias{coerce,ANY,NiftiMatrix-method}
\title{NIfTI images as DelayedArray objects, using HDF5Array}
\usage{
NiftiArray(filepath, name = "image", header_name = "hdr", type = NA,
  header = NULL)
}
\arguments{
\item{filepath}{The path (as a single character string) to the HDF5
file where the dataset is located.}

\item{name}{The name of the image in the HDF5 file.}

\item{header_name}{The name of the header in the HDF5 file.}

\item{type}{\code{NA} or the R atomic type, passed to
\code{\link[HDF5Array:HDF5Array]{HDF5Array::HDF5Array()}}.}

\item{header}{list of header information to override call of
\link{nifti_header}}
}
\value{
A \code{NiftiArray} object
}
\description{
The NiftiArray class is a \link[DelayedArray:DelayedArray]{DelayedArray::DelayedArray} subclass
represented in HDF5, with the header.
All the operations available for \link[HDF5Array:HDF5Array]{HDF5Array::HDF5Array}
objects work on NiftiArray objects.
}
\examples{
nii_fname = system.file("extdata", "example.nii.gz", package = "RNifti")
res = NiftiArray(nii_fname)
res2 = NiftiArray(slot(slot(res, "seed"), "filepath"))
}
