% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNiftiArray.R
\name{writeNiftiArray}
\alias{writeNiftiArray}
\title{Write NiftiArray Object}
\usage{
writeNiftiArray(x, filepath = tempfile(fileext = ".h5"),
  name = "image", header_name = "hdr", chunkdim = NULL,
  level = NULL, verbose = FALSE, header = NULL, overwrite = FALSE)
}
\arguments{
\item{x}{a \code{niftiImage} object or file path to \code{NIfTI} file}

\item{filepath}{The path (single character string) to the HDF5 file.}

\item{name}{The name of the image in the HDF5 file.}

\item{header_name}{The name of the header in the HDF5 file.}

\item{chunkdim}{The dimensions of the chunks to use for
writing the data to disk.
Passed to \link[HDF5Array:writeHDF5Array]{HDF5Array::writeHDF5Array}.}

\item{level}{The compression level,
passed to \link[HDF5Array:writeHDF5Array]{HDF5Array::writeHDF5Array}.}

\item{verbose}{Display progress,
passed to \link[HDF5Array:writeHDF5Array]{HDF5Array::writeHDF5Array}.}

\item{header}{list of header information;
overrides call of \link{nifti_header}}

\item{overwrite}{\code{FALSE} by default and an in the
event that an HDF5 file already exists for \code{filepath}
input then do not overwrite it.
If set to \code{TRUE} then the "image" and "hdr" objects at this file
location will overwrite.}
}
\value{
A \code{NiftiArray} object.
}
\description{
Write NiftiArray Object
}
\examples{
nii_fname = system.file("extdata", "example.nii.gz", package = "RNifti")
res = writeNiftiArray(nii_fname)
filepath = tempfile(fileext = ".h5")
res = writeNiftiArray(nii_fname, filepath = filepath)
testthat::expect_error(
   writeNiftiArray(nii_fname, filepath = filepath),
   regexp = "already exist",
)
res = writeNiftiArray(nii_fname, filepath = filepath, overwrite = TRUE)
img = RNifti::readNifti(nii_fname)
writeNiftiArray(c(img), header = nifti_header(img))
}
