% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNiftiArray.R
\name{writeNiftiArray}
\alias{writeNiftiArray}
\title{Write NiftiArray Object}
\usage{
writeNiftiArray(x, filepath = tempfile(fileext = ".h5"),
  name = "image", header_name = "hdr", chunkdim = NULL,
  level = NULL, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
Write NiftiArray Object
}
